Update on Mini HINT
Reworked some of SensorManager and SensorData.
Added new script named Sensor. This script is placed on any object in Mini HINT that you want to be a sensor. SensorManager will know which object contains the Sensor script. In this script the resource is added and if the sensor is lamp button or not.